enable repeated use of doxygen prepare script

+++ doc/doxygen/prepare_doxy_update.sh
@@ -4,17 +4,22 @@
 #    [path to source directory]/doc/doxygen prepare_doxy_update.sh
 #
 KV_SOURCE=$(kaliveda-config --srcdir)
-
-ln -s $KV_SOURCE kaliveda.git
+if [ ! -d ./kaliveda.git ]; then
+   ln -s $KV_SOURCE kaliveda.git
+fi
+if [ ! -d ./html ]; then
 git clone --single-branch -b gh-pages git@github.com:kaliveda-dev/kaliveda.git html
+fi
 
 # remove history from gh-pages
 cd html
 git reset --hard $(git rev-list --max-parents=0 --abbrev-commit HEAD)
 cd ..
 
-# copy scripts etc.
+# link scripts etc.
 list=$(ls $KV_SOURCE/doc/doxygen)
 for f in $list; do
-   ln -s $KV_SOURCE/doc/doxygen/$f
+   if [ ! -L $f ]; then
+      ln -s $KV_SOURCE/doc/doxygen/$f
+   fi
 done
